refactor: update namespaces to EntityMappings for EF configurations
