+++ src/Spoksy.Infrastructure/EntityMappings/MessageConfiguration.cs
@@ -2,7 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Spoksy.Domain.Entities;
 
-namespace Spoksy.Infrastructure.Configurations
+namespace Spoksy.Infrastructure.EntityMappings
 {
     public class MessageConfiguration : IEntityTypeConfiguration<Message>
     {
